fix: prevent download button layout shifts
